+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 ## Change log
 ----------------------
 
+Version 2.5
+-------------
+
+ADDED:
+
+- new test class for Period
+
+CHANGED:
+
+- removed deprecated methods from all classes
+- increase of code coverage
+
 Version 2.4
 -------------